Add docs and move to other.ts
