+++ src/index.ts
@@ -2,4 +2,3 @@ export * from './unions';
 export * from './input';
 export * from './storage';
 export * from './other';
-export * from './round-robin';
